Greate refactor
