+++ src/RTOV.ts
@@ -1,2 +1,1 @@
-export * from "./lib"
 export * from "./decorators"
