+++ src/lib/common.ts
@@ -1,0 +1,6 @@
+export interface MetaData {
+  className: string
+  schema: Object
+}
+
+export const isComplexType = (data: any) => typeof data === 'object';
