+++ src/lib/index.ts
@@ -1,8 +1,5 @@
-
-export interface MetaData {
-  className: string
-  schema: Object
-}
-
 export * from "./debug";
+export * from "./common"
+export * from "./prop-handlers"
 export * from "./array-setters";
+export * from './object-setters'
